// app/requests/_personal/postCreateBidFromTask/postCreateBidFromTaskSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PostCreateBidFromTaskSuccess {
  data: {
    id: number;
    user: {
      id: number;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name:
          | "admin"
          | "client"
          | "manager"
          | "recruiter"
          | "specialist"
          | "supervisor";
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    };
    status: 1 | 2 | 3 | 4 | 5;
    selfEmployed: boolean;
    place: {
      id: number;
      name: string;
      latitude: string;
      longitude: string;
      address_kladr: string;
      logo: string;
      region: {
        id: number;
        name: string;
        [k: string]: unknown;
      };
      brand: {
        id: number;
        name: string;
        logo: string;
        description: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    radius: null | number;
    price: null | number;
    priceResult: number;
    viewActivity: {
      id: number;
      name: string;
      detailName: string;
      previewText: string;
      logo: string;
      traveling: boolean;
      [k: string]: unknown;
    };
    dateStart: string;
    dateEnd: string;
    needFoto: boolean;
    dateActivity: {
      timeStart: string;
      timeEnd: string;
      places: {
        id: number;
        name: string;
        latitude: string;
        longitude: string;
        address_kladr: string;
        logo: string;
        region: {
          id: number;
          name: string;
          [k: string]: unknown;
        };
        brand: {
          id: number;
          name: string;
          logo: string;
          description: string;
          [k: string]: unknown;
        };
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    }[];
    order: null | {
      id: number;
      selfEmployed: boolean;
      status: number;
      user: {
        id: number;
        phone: number;
        email: string;
        logo: string;
        roles: {
          id: number;
          name:
            | "admin"
            | "client"
            | "manager"
            | "recruiter"
            | "specialist"
            | "supervisor";
          [k: string]: unknown;
        }[];
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    task: null | {
      id: number;
      selfEmployed: boolean;
      status: number;
      user: {
        id: number;
        phone: number;
        email: string;
        logo: string;
        roles: {
          id: number;
          name:
            | "admin"
            | "client"
            | "manager"
            | "recruiter"
            | "specialist"
            | "supervisor";
          [k: string]: unknown;
        }[];
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    acceptingUsers: {
      id: number;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name:
          | "admin"
          | "client"
          | "manager"
          | "recruiter"
          | "specialist"
          | "supervisor";
        [k: string]: unknown;
      }[];
      radius: string;
      name: string;
      age: string;
      country: string;
      viewActivities: string[];
      status: 1 | 2 | 3 | 4 | 5;
      [k: string]: unknown;
    }[];
    count: number;
    statistic: {
      accepted: number;
      count: number;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  };
  [k: string]: unknown;
}